Lexer: Support lexing and parsing `=%>` ERB closing tag (#568)
This pull request adds support to be able to lex and parse the `=%>` ERB
closing tag. Since it's use is quite unknown and not well-defined we
should be able to parse is, so we can guide and advice people in the
linter to now use it, i.e using the Right Trim rule introduced in #556.

Co-Authored-By: Domingo Edwards <dedwards@buk.cl>