Add `visitNode` and `visitERBNode` methods to improve `erb-right-trim` (#569)
This pull request introduces two new methods `visitNode(node: Node)` and
`visitERBNode(node: ERBNode)` in the JavaScript visitor that allows to
visit any node, or visit any ERB node. This is useful and allows us to
improve the `erb-right-trim` introduced in #556.

This now updates the `erb-right-trim` to also handle cases where the
right trimming is used when it has no effect (like on non-outputting ERB
Nodes like `<%`).

/cc @domingo2000