C: Make parser options a normal struct member of the parser (#625)
This pull request changes the way we store the parser options in the parser
struct.

Instead of a heap allocation, we just use a local member.

In addition to less pointer chasing when checking the options, we also
get an explicit definition of the default parser options.

Public interfaces are unchanged by this refactoring.