Engine: Support `<% =begin%>` block comments with no space (#712)
This pull request updates the
`herb_extract_ruby_to_buffer_with_semicolons` method, so we always put a
whitespace before the semicolon when extracting Ruby code from a
HTML+ERB template.

With that, we can support parsing, compiling, and rendering the
following template:

```erb
<%
=begin%>
  This, while unusual, is a legal form of commenting.
<%
=end%>
<div>Hey there</div>
```

The problem is that `=begin;` isn't valid Ruby syntax

```ruby

=begin;


=end;
 
```

Whereas this is:

```ruby

=begin ;


=end ;
 
```

So we always replace the closing `%>` ERB tag with `" ;"` instead of `";
"` so it always ends up with valid extracted Ruby code.

Addresses
https://github.com/marcoroth/herb/issues/562#issuecomment-3426871811

/cc @asilano