Playground: Ensure `track_whitespace` parser option persists (#713)
This pull request updates the `playground_controller` to make sure we
don't loose the `track_whitespace` on the deployed version at:
https://herb-tools.dev/playground

Since that one is embedded in an `<iframe>` we have to set the params on
`window.parent`. When running the playground locally `window.parent`
will refer to itself when there is no parent.