Engine: Add support for customizing escape functions (#720)
This pull request updates the `Herb::Engine` to accept new `attrfunc`,
`jsfunc`, and `cssfunc` options.

This enables ReActionView (and Rails) to customize these escape
functions (similar to how it's already customizing `escapefunc`) so that
we can avoid the double escaping.

Since ActionView uses `ActionView::OutputBuffer` instead of `String` it
would end up be double-escaped if we didn't allow it to be customized
and set to `""` (see
https://github.com/marcoroth/reactionview/issues/12).

Enables https://github.com/marcoroth/reactionview/pull/53