Introduce `@herb-tools/config` and `.herb.yml` configuration file (#747)
This pull request introduces a new `@herb-tools/config` package that
helps manage a new `.herb.yml` configuration file that can be used by
all the tools.

The Herb Linter and Herb Formatter are the only user-facing tools that
are going to be supported initially. This pull request also updates the
Language Server to fully respect this `.herb.yml` configuration file.

Additionally, it adds extensive integration for the Herb Visual Studio
Code Extension, so that using and updating the `.herb.yml` configuration
file is a really pleasant experience.



https://github.com/user-attachments/assets/dcffb7c2-bff6-4000-b92e-1c97a164acf7




Resolves #349 
Resolves #400
Resolves #589 
Resolves #709 
Resolves #744 

Enables #313
Enables #480