Language Server: Herb Linter Fix-on-save (#758)
This pull request introduces the functionality for the Herb Language
Server to be able to autofix Herb Linter offenses on save.

This builds on the new `--fix` option in the Herb Linter (introduced in
#622), which can automatically correct offenses without formatting the
rest of the document.

If both `Fix-on-Save` and `Format-on-Save` are active, the Language
Server will first fix the offenses and then format the document.


https://github.com/user-attachments/assets/070bd0a4-4551-4527-9b0c-01a3e3d44df3

The Visual Studio Code Extension also exposes a new "Fix on save"
toggle:

<img width="2016" height="679" alt="CleanShot 2025-11-01 at 02 15 50@2x"
src="https://github.com/user-attachments/assets/2fbfa731-d3ad-4b97-a953-c565c136dc7b"
/>