Introduce `@herb-tools/rewriter` package (#759)
This pull request introduces a new `@herb-tools/rewriter` package that
provides a plugin architecture for transforming HTML+ERB templates and
enables users to create custom transformations.

This pull request just introduces the `@herb-tools/rewriter` package and
the building blocks. These transformations can later be used for pre-
and post-format rewriters. It can also serve as a foundation for
re-implementing the linter autofixes using the rewriter architecture.

We also want use this architecture to integrate the
`@herb-tools/tailwind-class-sorter` package into the
`@herb-tools/formatter` so it can be run as part of the format-action.