Formatter: Skip formatting Scaffold Templates (#771)
This pull request adds a visitor to detect if a given ParseResult
contains escaped `<%%` or `<%%=` tags, which are typically used in Rails
Scaffold Templates.

In the case of a scaffold template we abort the formatting and just
return the original source, since we cannot reliably tell how to format
a scaffold template properly, yet.

Resolves https://github.com/marcoroth/herb/issues/673