Ruby: Add metadata and input to Ruby test snapshots (#785)
This pull request updates the format of the Ruby test snapshots to
contain metadata alongside the actual snapshot content.

Specifically, it adds YAML frontmatter to be able to store some
metadata, like the source input or the options passed. This makes it
easier to review the changed snapshots when you have all the needed
context in the same file. Otherwise you have to jump back and forth
between the snapshot files and the actual test file to see the original
input we are testing against.

Here's an example:

```js
---
source: "Parser::BooleanAttributesTest#test_0001_boolean attribute"
input: "<input required />"
---
@ DocumentNode (location: (1:0)-(1:18))
└── children: (1 item)
    └── @ HTMLElementNode (location: (1:0)-(1:18))
        ├── open_tag:
        │   └── @ HTMLOpenTagNode (location: (1:0)-(1:18))
        │       ├── tag_opening: "<" (location: (1:0)-(1:1))
        │       ├── tag_name: "input" (location: (1:1)-(1:6))
        │       ├── tag_closing: "/>" (location: (1:16)-(1:18))
        │       ├── children: (1 item)
        │       │   └── @ HTMLAttributeNode (location: (1:7)-(1:15))
        │       │       ├── name:
        │       │       │   └── @ HTMLAttributeNameNode (location: (1:7)-(1:15))
        │       │       │       └── children: (1 item)
        │       │       │           └── @ LiteralNode (location: (1:7)-(1:15))
        │       │       │               └── content: "required"
        │       │       │
        │       │       │
        │       │       ├── equals: ∅
        │       │       └── value: ∅
        │       │
        │       └── is_void: true
        │
        ├── tag_name: "input" (location: (1:1)-(1:6))
        ├── body: []
        ├── close_tag: ∅
        ├── is_void: true
        └── source: "HTML"
```