C: Make `hb_buffer_append_char` thread-safe (#793)
This pull request replaces the `static` temporary buffer in
`hb_buffer_append_char` with a stack-allocated buffer so parallel calls
no longer share mutable state and could corrupt appended output.

Before, it corrupted data like this on rare occasions:

```
  left: "<piv>\n                            \n    <p>                </p>\n           \n</div>"
 right: "<div>\n                            \n    <p>                </p>\n           \n</div>"
```

This was only noticeable in the Cargo tests, since they run so fast in
parallel and quick succession that it could happen occasionally.