Formatter: Preserve Frontmatter when formatting documents (#797)
This pull request adds support for preserving frontmatter when
formatting HTML+ERB documents.

When the first Node within the `DocumentNode` is a `HTMLTextNode` and
starts/ends with `---` we treat that `HTMLTextNode` as frontmatter
content and preserve/format it exactly as-is.

**Example**

The following document is now being preserved:

```erb
---
title: My Page
layout: application
published: true
---

<div class="container">
  <h1><%= @title %></h1>
</div>
```

Previously it was formatted as:

```erb
--- title: My Page layout: application published: true ---

<div class="container">
  <h1><%= @title %></h1>
</div>
```


Thanks to @joshuap for the nudge!