C: Use `hb_string_T` when initializing new tokens (#798)
This PR changes the signature of the `token_init` function to take
`hb_string_T` instead of a raw string.

The token still stores its value as a `char *`. This PR just changes the
interface for initializing tokens.

The call sites in the lexer are adapted to the new `token_init`
signature and buffer usages are minimized.

## Newly introduced functions

- `hb_string_range` a function that allows to get a substring from a
given string
- `hb_string_to_c_string_using_malloc` a version of
`hb_string_to_c_string` that doesn't rely on the arena allocator.
(Necessary since we don't have widespread use of the arena allocator
yet)