Parser: Re-parse on `unexpected ";"` error for more accurate errors (#902)
This pull request changes the way the parser handles the `unexpected
";"` errors.

In https://github.com/marcoroth/herb/pull/824, we changed the
`herb_extract_ruby_to_buffer_with_semicolons` method to only add
semicolons when needed. This pull request reverts that. But instead of
just outputting the `unexpected ";"` errors (as it was before #824), we
check if the `;` was actually part of the source file.

If it wasn't part of the source file we re-parse this individual
`ERBContentNode` to check for syntax errors in the raw source of that
`ERBContentNode` instead, which shouldn't have the semicolon and thus
shouldn't append the `unexpected ";"` errors.

Resolves https://github.com/marcoroth/herb/issues/854
Resolves https://github.com/marcoroth/herb/issues/865
Resolves #899