C: Compile with `-fvisibility=hidden` flag (#906)
This PR starts compiling the ruby herb gem c dependencies with
`-fvisibility=hidden`.

This addresses a criticial bug that lead to crashes when the ruby gem
ran in an environment that already had harfbuzz loaded. Both define a
function called `hb_buffer_append`, which resulted in conflicts.

With this fix we only expose the init function required by the ruby
runtime in the DLL/SO.

This PR addresses the issues reported in
https://github.com/marcoroth/herb/issues/905

## Before

![Bildschirmfoto_2025-11-19_um_00 09
11](https://github.com/user-attachments/assets/387d11b5-a62d-4548-9a38-7bfbd6ba808a)

## After

![Bildschirmfoto_2025-11-19_um_00 19
23](https://github.com/user-attachments/assets/5acfa368-351a-47a1-8b7e-20ac8351896f)