Formatter: Disable formatter in default config (#944)
This pull request changes the default configuration to not have the
formatter enabled by default. Since the formatter is still experimental
people should be opting in to use it.

Since the default configuration had the formatter enabled by default
this meant that the language server started formatting on save by
default as well, which is not ideal.

This pull request addresses this, so that people have to opt-in and
enable the formatter in teh `.herb.yml` configuration file.
Additionally, it splits up the fix and format concerns in the language
server, so that the `editor.formatOnSave` setting in VS Code is also
properly respected.

Resolves #910
Resolves #911